requirements.txt
python-telegram-bot==20.8
httpx==0.26.0
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
pandas==2.2.3
numpy==1.26.4
transformers==4.38.0
torch==2.2.0
sqlalchemy==2.0.27
apscheduler==3.10.4
beautifulsoup4>=4.12.0
lxml>=4.9.0
mplfinance==0.12.9b7
matplotlib>=3.7.0
telethon==1.34.0
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2